test: persistent cache
