+++ test/fixtures/empty.js
@@ -1,0 +1,1 @@
+export default 'bar';
